Reorganizing fetch utils and logger into different files
